requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.12.1
certifi==2023.11.17
click==8.1.7
cloudinary==1.37.0
Django==4.2.8
django-crispy-forms==2.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
factory-boy==3.3.0
faker==20.1.0
